requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
scipy>=1.9.0

# Time series analysis
statsmodels>=0.13.0
prophet>=1.1.0

# Machine learning
xgboost>=1.6.0
lightgbm>=3.3.0

# Deep learning (optional)
tensorflow>=2.10.0
torch>=1.12.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Data collection and APIs
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0
tqdm>=4.64.0

# Jupyter and notebooks
jupyter>=1.0.0
ipywidgets>=7.7.0

# Web dashboard (optional)
streamlit>=1.12.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
